Interesting note
Gameplay:
+Limit FPS to 30

It is worth noting... There is a lot tied to the framerate of the game and it's making it really weird between devices. For example ~ The speed of the NPCS, the animation rate of the dialogue box fade, the rotation speed of the interact button, even the jump speed of the player. Defo needs fixing.